move repository files into core folder
