+++ src/packages/core/repository/data-source/copy-data-source.interface.ts
@@ -1,4 +1,4 @@
-import type { DataSourceResponse } from '@umbraco-cms/backoffice/repository';
+import type { DataSourceResponse } from 'src/packages/core/repository';
 
 export interface UmbCopyDataSource {
 	copy(unique: string, targetUnique: string | null): Promise<DataSourceResponse<string>>;
