+++ src/packages/core/repository/data-source/item-data-source.interface.ts
@@ -1,4 +1,4 @@
-import type { DataSourceResponse } from '@umbraco-cms/backoffice/repository';
+import type { DataSourceResponse } from 'src/packages/core/repository';
 
 export interface UmbItemDataSource<ItemType> {
 	getItems(unique: Array<string>): Promise<DataSourceResponse<Array<ItemType>>>;
