+++ src/packages/core/repository/data-source/move-data-source.interface.ts
@@ -1,4 +1,4 @@
-import type { UmbDataSourceErrorResponse } from '@umbraco-cms/backoffice/repository';
+import type { UmbDataSourceErrorResponse } from 'src/packages/core/repository';
 
 export interface UmbMoveDataSource {
 	move(unique: string, targetUnique: string | null): Promise<UmbDataSourceErrorResponse>;
