+++ src/packages/core/repository/detail-repository.interface.ts
